Initial edits to convert. Split helper maps into convert_maps.
